Create Command abstract class for easy creation
